Support for ng-xi18n
add an i18n script in the package.json

This script will put files in the assets/i18n folder.

There is supporting code in zlux-app-manager/virtual-desktop
that will automatically look for a messages.${locale}.xlf file there to use
for translation.

Fixed various compile errors found by ng-xi18n (mostly
along the lines of member variables from the .ts file
that are referred to in the template can't be private).

Some trimming down of the tsconfig.i18n.json

Signed-off-by: Robert Penny <rpenny@rocketsoftware.com>